commenting out the bitwarden parts, to refactor them in their own file
